+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.5.4
 
 - added `WebHistory` and `WebHistoryItem` class
-- added `getCopyBackForwardList` method for `InAppWebView` and `InAppBrowser`
+- added `getCopyBackForwardList`, `goBackOrForward`, `canGoBackOrForward` and `goTo` methods for `InAppWebView` and `InAppBrowser`
 
 ## 0.5.3